bump docker

+++ src/github.com/docker/docker/pkg/progressreader/progressreader.go
@@ -1,3 +1,5 @@
+// Package progressreader provides a Reader with a progress bar that can be
+// printed out using the streamformatter package.
 package progressreader
 
 import (
@@ -7,30 +9,31 @@ import (
 	"github.com/docker/docker/pkg/streamformatter"
 )
 
-// Reader with progress bar
+// Config contains the configuration for a Reader with progress bar.
 type Config struct {
 	In         io.ReadCloser // Stream to read from
 	Out        io.Writer     // Where to send progress bar to
 	Formatter  *streamformatter.StreamFormatter
-	Size       int
-	Current    int
-	LastUpdate int
+	Size       int64
+	Current    int64
+	LastUpdate int64
 	NewLines   bool
 	ID         string
 	Action     string
 }
 
+// New creates a new Config.
 func New(newReader Config) *Config {
 	return &newReader
 }
 
 func (config *Config) Read(p []byte) (n int, err error) {
 	read, err := config.In.Read(p)
-	config.Current += read
-	updateEvery := 1024 * 512 //512kB
+	config.Current += int64(read)
+	updateEvery := int64(1024 * 512) //512kB
 	if config.Size > 0 {
 		// Update progress for every 1% read if 1% < 512kB
-		if increment := int(0.01 * float64(config.Size)); increment < updateEvery {
+		if increment := int64(0.01 * float64(config.Size)); increment < updateEvery {
 			updateEvery = increment
 		}
 	}
@@ -48,6 +51,7 @@ func (config *Config) Read(p []byte) (n int, err error) {
 	return read, err
 }
 
+// Close closes the reader (Config).
 func (config *Config) Close() error {
 	if config.Current < config.Size {
 		//print a full progress bar when closing prematurely
